+++ spec/cobrato/entities/payment_spec.rb
@@ -48,6 +48,10 @@ describe Cobrato::Entities::Payment do
       payment_option: "single_with_discount",
       renavam: "345352",
       uf: "rj",
+
+      state_registration: "2313",
+      active_debt_registration: "423413",
+      installment_number: 5
     }
   end
 
@@ -79,6 +83,9 @@ describe Cobrato::Entities::Payment do
       :gross_revenue, :gross_revenue_percentage,
 
       # IPVA specific
-      :city_code, :license_plate, :payment_option, :renavam, :uf
+      :city_code, :license_plate, :payment_option, :renavam, :uf,
+
+      # ICMS-SP specific
+      :state_registration, :active_debt_registration, :installment_number
     ]
 end