+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add DAS payment Support
 - Add IPVA payment Support
+- Add ICMS-SP payment Support
 
 ## v0.23.0.beta5